+++ client/components/CandidateReview/CandidateModals/ThankYouModal/ThankYouModal.css
