+++ client/components/CandidateReview/CandidateModals/ThankYouModal/index.jsx
