+++ client/components/CandidateReview/CandidateModals/common.css
